+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Event archive button.
 
+### Fixed
+- Typo label week electricity chart.
+
 ## [1.34.0] - 2022-09-30
 ### Added
 - Added solar to week chart.